Feature/refactor api tests (#15)
* Refactor Booking API tests: randomization, context structure, and data cleanup

* Refactor Booking API tests: randomization, context structure, and data cleanup

* Refactor Booking API tests: randomization, context structure, and data cleanup

* Refactor Booking API tests: randomization, context structure, and data cleanup

* refactored API commands and test data handling. Standardized options decomposition in requests, moved all static and dynamic test data to dedicated files, and improved error handling in the auth flow.

* refactored API commands and test data handling. Standardized options decomposition in requests, moved all static and dynamic test data to dedicated files, and improved error handling in the auth flow.

+++ cypress/support/api/booking.api.commands.js
@@ -1,51 +1,102 @@
 // cypress/support/api/booking.api.commands.js
-
-import { ENDPOINTS } from '../constants/api-urls';
-
 Cypress.Commands.add('booking__create', (bookingBody, options = {}) => {
+  const { failOnStatusCode, ...restOptions } = options;
+
   return cy.request({
+    ...restOptions,
     method: 'POST',
-    url: `${Cypress.env('apiBaseUrl')}${ENDPOINTS.BOOKING}`,
+    url: `${API_URLS.BASE_URL}${API_URLS.ENDPOINTS.BOOKING}`,
     headers: {
       'Content-Type': 'application/json',
       Accept: 'application/json',
+      ...restOptions.headers,
     },
     body: bookingBody,
-    ...options,
+    failOnStatusCode: failOnStatusCode,
+  });
+});
+
+Cypress.Commands.add('booking__getAllIds', (options = {}) => {
+  const { failOnStatusCode, ...restOptions } = options;
+
+  return cy.request({
+    ...restOptions,
+    method: 'GET',
+    url: `${API_URLS.BASE_URL}${API_URLS.ENDPOINTS.BOOKING}`,
+    failOnStatusCode: failOnStatusCode,
   });
 });
 
 Cypress.Commands.add('booking__getById', (bookingId, options = {}) => {
+  const { failOnStatusCode, ...restOptions } = options;
+
   return cy.request({
+    ...restOptions,
     method: 'GET',
-    url: `${Cypress.env('apiBaseUrl')}${ENDPOINTS.BOOKING}/${bookingId}`,
+    url: `${API_URLS.BASE_URL}${API_URLS.ENDPOINTS.BOOKING}/${bookingId}`,
     headers: {
       Accept: 'application/json',
+      ...restOptions.headers,
     },
-    ...options,
+    failOnStatusCode: failOnStatusCode,
   });
 });
 
-Cypress.Commands.add('booking__update', (bookingId, updateBody, token = Cypress.env('authToken')) => {
+Cypress.Commands.add('booking__fullUpdate', (bookingId, fullBody, options = {}) => {
+  const { failOnStatusCode, token: manualToken, ...restOptions } = options;
+  const token = manualToken || Cypress.env('token');
+
+  if (!token) throw new Error('Auth token is required for booking__fullUpdate');
   return cy.request({
-    method: 'PATCH',
-    url: `${Cypress.env('apiBaseUrl')}${ENDPOINTS.BOOKING}/${bookingId}`,
+    ...restOptions,
+    method: 'PUT',
+    url: `${API_URLS.BASE_URL}${API_URLS.ENDPOINTS.BOOKING}/${bookingId}`,
     headers: {
       'Content-Type': 'application/json',
       Accept: 'application/json',
       Cookie: `token=${token}`,
+      ...restOptions.headers,
+    },
+    body: fullBody,
+    failOnStatusCode: failOnStatusCode,
+  });
+});
+
+Cypress.Commands.add('booking__update', (bookingId, updateBody, options = {}) => {
+  const { failOnStatusCode, token: manualToken, ...restOptions } = options;
+  const token = manualToken || Cypress.env('token');
+
+  if (!token) throw new Error('Auth token is required for booking__update');
+  return cy.request({
+    ...restOptions,
+    method: 'PATCH',
+    url: `${API_URLS.BASE_URL}${API_URLS.ENDPOINTS.BOOKING}/${bookingId}`,
+    headers: {
+      'Content-Type': 'application/json',
+      Accept: 'application/json',
+      Cookie: `token=${token}`,
+      ...restOptions.headers,
     },
     body: updateBody,
+    failOnStatusCode: failOnStatusCode,
   });
 });
 
-Cypress.Commands.add('booking__delete', (bookingId, token = Cypress.env('authToken')) => {
+Cypress.Commands.add('booking__delete', (bookingId, options = {}) => {
+  const { failOnStatusCode, token: manualToken, ...restOptions } = options;
+  const token = manualToken || Cypress.env('token');
+
+  if (!token) throw new Error('Auth token is required for booking__delete');
+
   return cy.request({
+    ...restOptions,
     method: 'DELETE',
-    url: `${Cypress.env('apiBaseUrl')}${ENDPOINTS.BOOKING}/${bookingId}`,
+    url: `${API_URLS.BASE_URL}${API_URLS.ENDPOINTS.BOOKING}/${bookingId}`,
     headers: {
       'Content-Type': 'application/json',
       Cookie: `token=${token}`,
+      ...restOptions.headers,
     },
+    failOnStatusCode: failOnStatusCode,
   });
 });
